Aligned {add,with}_metaclass behavior with Python 3.X
Prior to this commit `add_metaclass` and `with_metaclass` could call the
`__new__` method of the metaclasses in the inheritance hierarchy multiple
times. For example, the following

    class AMeta(type):
        def __new__(cls, *args):
            print("AMeta.__new__")
            return super(AMeta, cls).__new__(cls, *args)

    class BMeta(AMeta):
        def __new__(cls, *args):
            print("BMeta.__new__")
            return super(BMeta, cls).__new__(cls, *args)

    @six.add_metaclass(BMeta)
    class Abstract(object):
        pass

    @six.add_metaclass(AMeta)
    class Concrete(Abstract):
        pass

results in

    BMeta.__new__
    AMeta.__new__
    type.__new__

while constructing `Abstract`, but

    AMeta.__new__
    type.__new__
    BMeta.__new__
    AMeta.__new__
    type.__new__

for `Concrete`. After this commit, both class definitions produce the
same chain of `__new__` calls.

The corresponding fix in 3.X is (I think) python/cpython@de31b191.